djangoapi/pyproject.toml
[tool.poetry]
name = "api-appointments"
version = "0.1.0"
description = "API REST para gestão de profissionais de saúde e consultas"
authors = ["viviane santos <vjsantos.viviane@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
django = ">=6.0,<7.0"
djangorestframework = ">=3.16.1"
psycopg2-binary = ">=2.9.11"
python-decouple = ">=3.8"
django-cors-headers = ">=4.9.0"
drf-spectacular = ">=0.29.0"
drf-spectacular-sidecar = ">=2025.12.1"
django-filter = ">=25.2"
djangorestframework-simplejwt = ">=5.5.1"
dotenv = "^0.9.9"
flake8 = "^7.3.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2"
pytest-django = ">=4.11.1"
pylint = ">=4.0.4"
black = ">=25.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
